Split MadTank into Attack* traits and WithChargeAnimation.
